+++ packages/twopoint5d/src/stage/Canvas2DStage.ts
@@ -1,6 +1,6 @@
 import {emit, eventize} from '@spearwolf/eventize';
 import {Sprite, SpriteMaterial, Texture, type Scene} from 'three';
-import type {ThreeRendererType} from '../display/types.js';
+import type {DisplayRendererType} from '../display/types.js';
 import {TextureFactory} from '../texture/TextureFactory.js';
 import {OrthographicProjection} from './OrthographicProjection.js';
 import {Stage2D} from './Stage2D.js';
@@ -9,7 +9,7 @@ import {StageRenderer} from './StageRenderer.js';
 export type Canvas2DStageFitType = 'contain' | 'cover';
 
 export class Canvas2DStage {
-  readonly renderer: ThreeRendererType;
+  readonly renderer: DisplayRendererType;
   readonly stageRenderer: StageRenderer;
 
   #fit: Canvas2DStageFitType = 'contain';
@@ -56,7 +56,7 @@ export class Canvas2DStage {
   #lastHeight = 0;
 
   constructor(
-    renderer: ThreeRendererType,
+    renderer: DisplayRendererType,
     ...args:
       | [width: number, height: number]
       | [width: number, height: number, fit: Canvas2DStageFitType]